Refactor: Add loading spinner to update page
Added a loading spinner to the update page in order to provide visual feedback to the user while the version badge is being fetched. This allows for a better user experience and prevents confusion when the badge takes longer to load. The spinner is now independent of the popup and can be imported by any viewer without loading the entire popup. Updated the HTML file to include the necessary CSS file for the spinner.